requirements.txt
Flask>=3.0.0
opencv-python
numpy
Pillow
onnxruntime
torch
torchvision
basicsr
facexlib
gfpgan
realesrgan